requirements.txt
pandas==1.3.3
python-dotenv==0.19.0
mypy==0.910
gspread==4.0.1
data-science-types==0.2.23
pytest==6.2.5